Move dx test into own file
